+++ internal/photoprism/purge.go
@@ -247,28 +247,39 @@ func (w *Purge) Start(opt PurgeOptions) (purgedFiles map[string]bool, purgedPhot
 
 	log.Info("purge: searching index for hidden media files")
 
+	// Set photo quality scores to -1 if files are missing.
 	if err := query.ResetPhotoQuality(); err != nil {
 		return purgedFiles, purgedPhotos, err
 	}
 
-	if err := query.PurgeOrphans(); err != nil {
-		log.Errorf("purge: %s (orphans)", err)
+	// Remove orphan index entries.
+	if opt.Dry {
+		if files, err := query.OrphanFiles(); err != nil {
+			log.Errorf("purge: %s (find orphan files)", err)
+		} else if l := len(files); l > 0 {
+			log.Infof("purge: found %d orphan files", l)
+		} else {
+			log.Infof("purge: found no orphan files")
+		}
+	} else {
+		if err := query.PurgeOrphans(); err != nil {
+			log.Errorf("purge: %s (remove orphans)", err)
+		}
 	}
 
+	// Hide missing album contents.
 	if err := query.UpdateMissingAlbumEntries(); err != nil {
 		log.Errorf("purge: %s (album entries)", err)
 	}
 
-	log.Info("purge: updating photo counts")
-
 	// Update precalculated photo and file counts.
+	log.Info("purge: updating photo counts")
 	if err := entity.UpdatePhotoCounts(); err != nil {
 		log.Errorf("purge: %s (update counts)", err)
 	}
 
-	log.Info("purge: updating preview images")
-
 	// Update album, subject, and label preview thumbs.
+	log.Info("purge: updating preview images")
 	if err := query.UpdatePreviews(); err != nil {
 		log.Errorf("purge: %s (update previews)", err)
 	}
